pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ddl2data"
version = "0.3.0"
description = "Turn any SQL schema into realistic test data — instantly."
readme = "README.md"
license = "Apache-2.0"
license-files = ["LICENSE"]
requires-python = ">=3.10"
keywords = ["synthetic-data", "sql", "ddl", "cli", "faker", "test-data", "schema", "ddl2data"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Database",
  "Topic :: Software Development :: Testing",
  "Topic :: Utilities",
]
dependencies = [
  "sqlglot>=25.0.0",
  "Faker>=25.0.0",
  "SQLAlchemy>=2.0.0",
  "PyYAML>=6.0.0",
  "tomli>=2.0.1; python_version < '3.11'",
]

[project.urls]
Homepage = "https://github.com/hyangminj/ddl2data"
Repository = "https://github.com/hyangminj/ddl2data"
Issues = "https://github.com/hyangminj/ddl2data/issues"
Releases = "https://github.com/hyangminj/ddl2data/releases"

[project.optional-dependencies]
polars = [
  "polars>=1.0.0",
]
test = [
  "pytest>=8.0.0",
  "psycopg2-binary>=2.9",
  "boto3>=1.34",
  "google-cloud-bigquery>=3.0.0",
]

[project.scripts]
ddl2data = "ddl2data.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
  "integration: requires external services or credentials",
  "postgres: requires PostgreSQL test database",
  "dynamodb: requires LocalStack DynamoDB test service",
  "bigquery: requires real Google BigQuery project credentials",
]
